Fix echo example to be more robust
The geolocation service can have ratelimits.  Support using
`$GEOLOCATION_JSON_URL` from environment to override the URL.  As long as the
schema of what's returned is the same, this will work.  This has been tested to
work well with a paid license for the same service we use by default, which
adds `?key=SECRET` to the URL.

Add handling for in-response errors, for improved diagnostics.
Increase timeout for the API service, based on past observed behavior.
Log our geolocation handling at startup (while censoring as needed).